Cleaned code to compile with -Wextra
Added -Wextra to Makefile and corrected warnings that were emitted by
GCC. Also tried to add -Wpedantic, but some warnings emitted by it could
not be removed (due to macro usage of min and max using GCC extensions).

+++ source/common.h
@@ -44,14 +44,14 @@
 
 inline char* strupper(const char* str) {
     char* buffer = (char*)malloc(strlen(str) + 1);
-    for (int i = 0; i < strlen(str); ++i)
+    for (size_t i = 0; i < strlen(str); ++i)
         buffer[i] = toupper((unsigned)str[i]);
     return buffer;
 }
 
 inline char* strlower(const char* str) {
     char* buffer = (char*)malloc(strlen(str) + 1);
-    for (int i = 0; i < strlen(str); ++i)
+    for (size_t i = 0; i < strlen(str); ++i)
         buffer[i] = tolower((unsigned)str[i]);
     return buffer;
 }
